refactor: Http 관련 유틸 util.http 패키지로 묶음
