+++ src/main/java/controller/ResourceController.java
@@ -1,6 +1,9 @@
 package controller;
 
-import util.*;
+import util.http.ContentType;
+import util.http.HttpRequest;
+import util.http.HttpResponse;
+import util.http.HttpStatus;
 
 import java.io.File;
 import java.io.IOException;
