+++ src/main/java/util/RequestParser.java
@@ -1,5 +1,7 @@
 package util;
 
+import util.http.HttpRequest;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
