+++ src/main/java/util/http/HttpResponse.java
@@ -1,4 +1,4 @@
-package util;
+package util.http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
