+++ src/main/java/util/http/HttpStatus.java
@@ -1,4 +1,4 @@
-package util;
+package util.http;
 
 public enum HttpStatus {
