+++ src/main/java/webserver/RequestHandler.java
@@ -7,6 +7,8 @@ import java.util.function.BiConsumer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import util.*;
+import util.http.HttpRequest;
+import util.http.HttpResponse;
 
 public class RequestHandler implements Runnable {
